+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.4
+- fix getSec read from raw
+
+## 0.0.3
+- make `SocketOut().add` private
+- make `SocketOut().getLength` static
+
+## 0.0.2
+- fix: mutex lock error
+- update tests
+
 ## 0.0.1
 
 * TODO: Describe initial release.